jobui/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
boto3==1.35.0
aiofiles==24.1.0
pydantic-settings==2.5.0
